# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ConsultMessage": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "read_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "sender_physician_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender Physician Id"
          },
          "thread_id": {
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "thread_id",
          "body",
          "created_at"
        ],
        "title": "ConsultMessage",
        "type": "object"
      },
      "ConsultThread": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "initiator_physician_id": {
            "title": "Initiator Physician Id",
            "type": "string"
          },
          "last_message": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConsultMessage"
              },
              {
                "type": "null"
              }
            ]
          },
          "recipient_physician_id": {
            "title": "Recipient Physician Id",
            "type": "string"
          },
          "referral_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Id"
          },
          "unread_count": {
            "default": 0,
            "title": "Unread Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "initiator_physician_id",
          "recipient_physician_id",
          "created_at"
        ],
        "title": "ConsultThread",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Patient": {
        "description": "Synthetic patient. Never put real PHI here.",
        "properties": {
          "age": {
            "title": "Age",
            "type": "integer"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "insurance": {
            "$ref": "#/components/schemas/InsurancePlan"
          },
          "location": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "zip_code": {
            "title": "Zip Code",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "age",
          "zip_code",
          "location",
          "insurance"
        ],
        "title": "Patient",
        "type": "object"
      },
...
      "Referral": {
        "properties": {
          "appointment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Appointment"
              },
              {
                "type": "null"
              }
            ]
          },
          "approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Approval"
              },
              {
                "type": "null"
              }
            ]
          },
          "case_notes": {
            "title": "Case Notes",
            "type": "string"
          },
          "complexity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Complexity"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "parsed_case": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParsedCase"
              },
              {
                "type": "null"
              }
            ]
          },
          "patient_id": {
            "title": "Patient Id",
            "type": "string"
          },
          "referring_physician_id": {
            "title": "Referring Physician Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ReferralStatus"
          },
          "trial_approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrialApproval"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "urgency": {
            "$ref": "#/components/schemas/Urgency"
          }
        },
        "required": [
          "id",
          "patient_id",
          "referring_physician_id",
          "case_notes",
          "urgency",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "Referral",
        "type": "object"
      },
...
      "SendMessageRequest": {
        "properties": {
          "body": {
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "sender_physician_id": {
            "title": "Sender Physician Id",
            "type": "string"
          }
        },
        "required": [
          "sender_physician_id",
          "body"
        ],
        "title": "SendMessageRequest",
        "type": "object"
      },
...
      "StartConsultRequest": {
        "properties": {
          "body": {
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "from_physician_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Physician Id"
          },
          "referral_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Id"
          },
          "to_physician_id": {
            "title": "To Physician Id",
            "type": "string"
          }
        },
        "required": [
          "to_physician_id",
          "body"
        ],
        "title": "StartConsultRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "Dev mode only (AUTH_MODE=dev): act as this physician id, e.g. doc_001",
        "in": "header",
        "name": "X-Dev-Physician",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "description": "Auth0 access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Physician-first referral matching. Nothing is sent or booked without physician approval.",
    "title": "ECHO API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/consults": {
      "get": {
        "description": "The signed-in physician's threads, each with its last message and unread count.\n`physician_id` is optional; if sent it must be the signed-in physician.",
        "operationId": "list_consults_consults_get",
        "parameters": [
          {
            "in": "query",
            "name": "physician_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Physician Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsultThread"
                  },
                  "title": "Response List Consults Consults Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Consults",
        "tags": [
          "consult"
        ]
      },
      "post": {
        "operationId": "start_consult_consults_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartConsultRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultThread"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Start Consult",
        "tags": [
          "consult"
        ]
      }
    },
    "/consults/{thread_id}/messages": {
      "get": {
        "description": "Full history; marks the other participant's messages as read by the signed-in physician.\n`reader_physician_id` is optional; if sent it must be the signed-in physician.",
        "operationId": "get_consult_messages_consults__thread_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "reader_physician_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reader Physician Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsultMessage"
                  },
                  "title": "Response Get Consult Messages Consults  Thread Id  Messages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Consult Messages",
        "tags": [
          "consult"
        ]
      },
      "post": {
        "operationId": "send_consult_message_consults__thread_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultMessage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Send Consult Message",
        "tags": [
          "consult"
        ]
      }
    },
...
    "/patients": {
      "get": {
        "operationId": "list_patients_patients_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  },
                  "title": "Response List Patients Patients Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Patients",
        "tags": [
          "directory"
        ]
      }
    },
...
    "/referrals": {
      "get": {
        "description": "The signed-in physician's referrals (an admin sees everyone's).",
        "operationId": "list_referrals_referrals_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Referral"
                  },
                  "title": "Response List Referrals Referrals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Referrals",
        "tags": [
          "referrals"
        ]
      },
      "post": {
        "operationId": "create_referral_referrals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferralCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Referral"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Referral",
        "tags": [
          "referrals"
        ]
      }
    },
    "/referrals/{referral_id}": {
      "get": {
        "operationId": "get_referral_referrals__referral_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "referral_id",
            "required": true,
            "schema": {
              "title": "Referral Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Referral"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Referral",
        "tags": [
          "referrals"
        ]
      },
      "patch": {
        "description": "Physician confirms or overrides urgency and complexity.",
        "operationId": "confirm_referral_referrals__referral_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "referral_id",
            "required": true,
            "schema": {
              "title": "Referral Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferralUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Referral"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Confirm Referral",
        "tags": [
          "referrals"
        ]
      }
    },
    "/referrals/{referral_id}/approve": {
      "post": {
        "description": "The physician's explicit approval. This is the only path that approves or books.",
        "operationId": "approve_and_book_referrals__referral_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "referral_id",
            "required": true,
            "schema": {
              "title": "Referral Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Referral"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "Approver is not the referring physician"
          },
          "409": {
            "description": "Stale match run, or slot no longer available; re-run matching"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Approve And Book",
        "tags": [
          "referrals"
        ]
      }
    },
...
    "/referrals/{referral_id}/trials": {
      "get": {
        "operationId": "list_trials_referrals__referral_id__trials_get",
        "parameters": [
          {
            "in": "path",
            "name": "referral_id",
            "required": true,
            "schema": {
              "title": "Referral Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Trial"
                  },
                  "title": "Response List Trials Referrals  Referral Id  Trials Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Not parsed, or complexity is not rare_complex"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Trials",
        "tags": [
          "trials"
        ]
      }
    },
...
  }
}